pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polymarket-ai-trader"
version = "0.1.0"
description = "AI-powered probability trading system for Polymarket"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    # Trading
    "py-clob-client>=0.1.0",
    "web3>=6.11.0",
    "aiohttp>=3.9.0",
    # Machine Learning
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.1.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    # NLP
    "transformers>=4.35.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.1.0",
    # Vector Database
    "chromadb>=0.4.0",
    # Data Sources
    "newsapi-python>=0.2.7",
    "tweepy>=4.14.0",
    "praw>=7.7.0",
    "httpx>=0.25.0",
    # Infrastructure
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "celery>=5.3.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    # Monitoring
    "mlflow>=2.9.0",
    "prometheus-client>=0.19.0",
    # Utils
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "structlog>=23.2.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

